Small Elements and draft for the Arf Closure of a good semigroup
I added a new function that computes the small elements of an Arf
semigroup. I also added a function that computes the Arf Closure of an
Arf good semigroup. The input must be the set of small elements of the
Arf semigroup, but it is still missing a function that checks the input
( at the moment it only checks if there is a conductor).