Update GoodsemigroupsGAP.g
I created a unique function to compute the optimized bound "BoundForConductorOfGoodSemigroupsContaining". It contains:
BoundNaive <-> The function that compute the first bound
BoundRefine <-> The function the refine the bound computed from the previous one
FiniteSet <-> It is a function that, from a set of vectors containing infinities, produces a new set of vector containing the same minimal good semigroup but having only finite elements (the old Minimizza).
I tried to adjust the alignments, I add new comments and changed some of the old, I renamed some variables.